# api/gen/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/users/me/": {
            "get": {
                "summary": "Get My User",
                "operationId": "get_my_user_users_me__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticatedUser"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/users/{user_id}/": {
            "get": {
                "summary": "Get User By Id",
                "operationId": "get_user_by_id_users__user_id___get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "User Id",
                            "type": "integer"
                        },
                        "name": "user_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recipes/": {
            "get": {
                "summary": "Get Recipes",
                "operationId": "get_recipes_recipes__get",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Page",
                            "type": "integer",
                            "default": 1
                        },
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Per Page",
                            "type": "integer",
                            "default": 10
                        },
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedRecipes"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            },
            "post": {
                "summary": "Create User Recipe",
                "operationId": "create_user_recipe_recipes__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/recipes/{recipe_id}/": {
            "get": {
                "summary": "Get Single Recipe",
                "operationId": "get_single_recipe_recipes__recipe_id___get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Recipe Id",
                            "type": "integer"
                        },
                        "name": "recipe_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            },
            "post": {
                "summary": "Update Recipe",
                "operationId": "update_recipe_recipes__recipe_id___post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Recipe Id",
                            "type": "integer"
                        },
                        "name": "recipe_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeEdit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeInDB"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/recipes/{recipe_id}/generate-pdf/": {
            "get": {
                "summary": "Generate Recipe Pdf",
                "operationId": "generate_recipe_pdf_recipes__recipe_id__generate_pdf__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Recipe Id",
                            "type": "integer"
                        },
                        "name": "recipe_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/users/{author_id}/recipes/": {
            "get": {
                "summary": "Get User Recipes",
                "operationId": "get_user_recipes_users__author_id__recipes__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Author Id",
                            "type": "integer"
                        },
                        "name": "author_id",
                        "in": "path"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Page",
                            "type": "integer",
                            "default": 1
                        },
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Per Page",
                            "type": "integer",
                            "default": 10
                        },
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedRecipes"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/users/{author_id}/recipes/generate-pdf/": {
            "get": {
                "summary": "Generate User Recipes Pdf",
                "operationId": "generate_user_recipes_pdf_users__author_id__recipes_generate_pdf__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Author Id",
                            "type": "integer"
                        },
                        "name": "author_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/recipes/ingredients/{ingredient_id}/": {
            "post": {
                "summary": "Update Ingredient",
                "operationId": "update_ingredient_recipes_ingredients__ingredient_id___post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Ingredient Id",
                            "type": "integer"
                        },
                        "name": "ingredient_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeIngredientEdit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Update Ingredient Recipes Ingredients  Ingredient Id   Post",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RecipeIngredientInDB"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/recipes/{recipe_id}/ingredients/": {
            "post": {
                "summary": "Add Ingredient",
                "operationId": "add_ingredient_recipes__recipe_id__ingredients__post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Recipe Id",
                            "type": "integer"
                        },
                        "name": "recipe_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeIngredientCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Add Ingredient Recipes  Recipe Id  Ingredients  Post",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RecipeIngredientInDB"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
        "/recipes/steps/{step_id}/": {
            "post": {
                "summary": "Update Step",
                "operationId": "update_step_recipes_steps__step_id___post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Step Id",
                            "type": "integer"
                        },
                        "name": "step_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecipeStepEdit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Update Step Recipes Steps  Step Id   Post",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RecipeStepInDB"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearerWithCookie": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthenticatedUser": {
                "title": "AuthenticatedUser",
                "required": [
                    "email",
                    "id",
                    "token",
                    "role"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "format": "email"
                    },
                    "first_name": {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "token": {
                        "$ref": "#/components/schemas/Token"
                    },
                    "role": {
                        "title": "Role",
                        "type": "string"
                    }
                }
            },
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "PaginatedRecipes": {
                "title": "PaginatedRecipes",
                "required": [
                    "page",
                    "per_page",
                    "max_page",
                    "result_count",
                    "data"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "title": "Page",
                        "type": "integer"
                    },
                    "per_page": {
                        "title": "Per Page",
                        "type": "integer"
                    },
                    "max_page": {
                        "title": "Max Page",
                        "type": "integer"
                    },
                    "result_count": {
                        "title": "Result Count",
                        "type": "integer"
                    },
                    "data": {
                        "title": "Data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeInDB"
                        }
                    }
                }
            },
...
            "User": {
                "title": "User",
                "required": [
                    "email",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "format": "email"
                    },
                    "first_name": {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearerWithCookie": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "auth/token/"
                    }
                }
            }
        }
    }
}